% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{allAvailableSeries}
\alias{allAvailableSeries}
\title{List All Data Series Available for Download}
\usage{
allAvailableSeries()
}
\value{
Character vector of series names.
}
\description{
List All Data Series Available for Download
}
\examples{

print(allAvailableSeries())

}
\author{
Jasper Watson
}
