% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B10}
\alias{B10}
\title{TWI Weights}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/trade-weighted-index-weights-and-trade-volumes} for a full description.
}
\keyword{internal}
