% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B13}
\alias{B13}
\title{B13 - Historical TWI weights}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b13} for a full description.
}
\note{
The historical data is split across multiple files at source. We download all
      of them and merge together to form a complete history.
}
