% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B20}
\alias{B20}
\title{New Residential Mortgage Standard Interest Rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/new-residential-mortgage-standard-interest-rates}
for a full description.
}
\keyword{internal}
