% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B21}
\alias{B21}
\title{B21 - New residential mortgage special interest rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b21-new-residential-mortgage-special-interest-rates}
for a full description.
}
\keyword{internal}
