% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B25}
\alias{B25}
\title{New Interest Bearing Call Savings Account Interest Rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b25-new-interest-bearing-call-savings-account-interest-rates}
for a full description.
}
\keyword{internal}
