% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B27}
\alias{B27}
\title{New Interest Bearing Term PIE Deposits Interest Rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/new-interest-bearing-term-portfolio-investment-entity-deposit-interest-rates}
for a full description.
}
\keyword{internal}
