% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B3}
\alias{B3}
\title{Retail Interest Rates on Lending and Deposits}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/retail-interest-rates-on-lending-and-deposits} for a full description.
}
\keyword{internal}
