% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B30}
\alias{B30}
\title{New Residential Mortgage Weighted Average Interest Rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/new-residential-mortgage-weighted-average-interest-rates}
for a full description.
}
\keyword{internal}
