% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B4}
\alias{B4}
\title{B4 - Foreign exchange turnover ($m daily averages)}
\arguments{
\item{option}{Allowed values are "NZD" (default), "USD" and "EUR".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b4} for a full description.
}
