% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B7}
\alias{B7}
\title{Costs of Deposits}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/exchange-and-interest-rates/costs-of-deposits} for a full
description.
}
\keyword{internal}
