% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C13}
\alias{C13}
\title{Credit Card Spending}
\arguments{
\item{option}{Allowed values are \dQuote{issued} (default) and \dQuote{used}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/credit-card-spending} for a full description.
}
\keyword{internal}
