% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C21}
\alias{C21}
\title{Key Household Financial Statistics}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/households/key-household-financial-statistics} for a full description.
}
\keyword{internal}
