% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C22}
\alias{C22}
\title{Household Balance Sheet}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/households/household-balance-sheet} for a full description.
}
\keyword{internal}
