% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C30}
\alias{C30}
\title{New Residential Mortgage Lending by Loan-To-Valuation Ratio}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/new-residential-mortgage-lending-by-loan-to-valuation-ratio} for a full description.
}
\keyword{internal}
