% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C33}
\alias{C33}
\title{New Residential Mortgage Lending By Purpose}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/new-residential-mortgage-lending-by-purpose} for a full description.
}
\keyword{internal}
