% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C5}
\alias{C5}
\title{Sector Lending}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/registered-banks-and-non-bank-lending-institutions-sector-lending} for a full description.
}
\keyword{internal}
