% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C50}
\alias{C50}
\title{C50 - Money and credit aggregates (depository corporations)}
\arguments{
\item{option}{Allowed values are "default" (default) and "longrun".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c50-money-and-credit-aggregates}
for a full description.
}
\keyword{internal}
