% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C51}
\alias{C51}
\title{Other Depository Corporations Analytical Accounts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c51-other-depository-corporations-analytical-accounts} for a full description.
}
\keyword{internal}
