% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C52}
\alias{C52}
\title{Depository Corporations Analytical Accounts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/depository-corporations-analytical-accounts} for a full description.
}
\keyword{internal}
