% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C55}
\alias{C55}
\title{Other Financial Corporations Analytical Accounts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/other-financial-corporations-analytical-accounts} for a full description.
}
\keyword{internal}
