% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C60}
\alias{C60}
\title{C60 - Credit conditions survey}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c60-credit-conditions-survey} for a full description.
}
\note{
The historical data is split across multiple files at source. We download all
      of them and merge together to form a complete history.
}
