% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C65}
\alias{C65}
\title{Bank Customer Lending Flows}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/bank-customer-lending-flows}
for a full description.
}
\keyword{internal}
