% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C70}
\alias{C70}
\title{New Lending By Purpose}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/new-lending-by-purpose}
for a full description.
}
\keyword{internal}
