% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C71}
\alias{C71}
\title{New Lending Fully Secured By Residential Mortgage}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/lending-and-monetary/new-lending-fully-secured-by-residential-mortgage}
for a full description.
}
\keyword{internal}
