% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D12}
\alias{D12}
\title{Standing Facilities}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/reserve-bank/standing-facilities} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
