% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D3}
\alias{D3}
\title{Open Market Operations}
\description{
See \url{https://www.rbnz.govt.nz/statistics/d3} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
