% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D31}
\alias{D31}
\title{D31 - Non-resident holdings of individual bonds (nominal value)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/d31} for a full description.
}
\keyword{internal}
