% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D35}
\alias{D35}
\title{Holdings of Kauri Bonds (Nominal Value)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/new-zealand-debt-securities/holdings-of-kauri-bonds} for a full description.
}
\keyword{internal}
