% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{F3}
\alias{F3}
\title{F3 - Bank notes in the hands of the public}
\description{
See \url{https://www.rbnz.govt.nz/statistics/f3} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
