% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{F4}
\alias{F4}
\title{Coin Mintings}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/reserve-bank/coin-mintings} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
