% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{F5}
\alias{F5}
\title{F5 - Reserve Bank foreign currency assets and liabilities}
\description{
See \url{https://www.rbnz.govt.nz/statistics/f5} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
