% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDownloadLink.R
\name{getDownloadLink}
\alias{getDownloadLink}
\title{Determine the URL for a File to be Downloaded}
\usage{
getDownloadLink(subFileName, seriesPage, series)
}
\arguments{
\item{subFileName}{The name of the file to be downloaded, as provided by
seriesOptions(). May contain only part of a series.}

\item{seriesPage}{An XML document returned by xml2::read_html containing the
webpage pertaining to the desired series.}

\item{series}{Character string giving the name of the data series to retrieve.}
}
\value{
Character string giving the URL of the file to be downloaded.
}
\description{
Determine the URL for a File to be Downloaded
}
\author{
Jasper Watson
}
\keyword{internal}
