% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReleaseCalendar.R
\name{getReleaseCalendar}
\alias{getReleaseCalendar}
\title{Download the Release Calendar for RBNZ Datasets}
\usage{
getReleaseCalendar(excludeDailyReleases = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{excludeDailyReleases}{Logical indicating whether to exclude series that
released daily from the calendar returned.}

\item{verbose}{Logical to be passed to RSelenium::rsDriver and/or
wdman::phantomjs. Defaults to FALSE}

\item{...}{Other arguments to pass to RSelenium::rsDriver. If one of these
arguments is \dQuote{browser} then the function will use that
instead of PhantomJS, see the details and examples below.}
}
\value{
Data frame containing upcoming release dates.
}
\description{
Downloads the table of release dates the RBNZ provides at
\url{https://www.rbnz.govt.nz/statistics/statistics-release-calendar}.
}
\details{
The release calendar on the RBNZ website requires a few button clicks
         to display so we use the RSelenium package to trigger the necessary
         javascript.
         
         The default option uses PhantomJS \url{https://phantomjs.org/}.
         This needs to be installed and on your path.
         
         Alternatively one can also use a headless Firefox or Chrome browser.
         The former requires geckodriver to be installed and on your path,
         \url{https://github.com/mozilla/geckodriver}, while the latter requires
         ChromeDriver, \url{https://sites.google.com/a/chromium.org/chromedriver/}.
         These both require a flag to be passed to RSelenium::rsDriver to signal
         that a headless browser should be used, see the examples below.
}
\examples{

\dontrun{

## The default option uses PhantomJS:
x <- getReleaseCalendar()

## Instead of PhantomJS one could also use firefox:
x <- getReleaseCalendar(browser = 'firefox',
                        extraCapabilities = list("moz:firefoxOptions" =
                                              list(args = list('--headless')))
                        )

## Or Chrome:
x <- getReleaseCalendar(browser = 'chrome',
                        extraCapabilities = list("chromeOptions" =
                                              list(args = list('--headless')))
                        )
## Sometimes you might have to specify the chromever argument;
## see RSelenium::rsDriver for details.

}

}
\author{
Jasper Watson
}
