% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeriesPage.R
\name{getSeriesPageURL}
\alias{getSeriesPageURL}
\title{Determine the URL of a Webpage Pertaining to a Series}
\usage{
getSeriesPageURL(series)
}
\arguments{
\item{series}{Character string giving the name of the data series to retrieve.}
}
\value{
Character string giving the URL of the series webpage.
}
\description{
Deprecated - the series pages are no longer used.
}
\author{
Jasper Watson
}
\keyword{internal}
