% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{H1}
\alias{H1}
\title{Household Inflation Expectations}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/households/household-inflation-expectations} for
a full description.
}
\keyword{internal}
