% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{J10}
\alias{J10}
\title{Insurance: Income Statement}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/insurance-financial-statistics/insurance-income-statement}
for a full description.
}
\keyword{internal}
