% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{J20}
\alias{J20}
\title{Insurance: Balance Sheet}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/insurance-financial-statistics/insurance-balance-sheet}
for a full description.
}
\keyword{internal}
