% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{L2}
\alias{L2}
\title{Core Funding Ratio}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-core-funding-ratio} for a full description.
}
\keyword{internal}
