% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{L3}
\alias{L3}
\title{Face Value of Funding by Residual Maturity ($m)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-face-value-of-funding-by-residual-maturity} for a full description.
}
\keyword{internal}
