% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{M15}
\alias{M15}
\title{Business Expectations Survey}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/economic-indicators/business-expectations-survey} for a full description.
}
\keyword{internal}
