% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{M5}
\alias{M5}
\title{Gross Domestic Product}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/economic-indicators/gross-domestic-product} for a full description.
}
\keyword{internal}
