% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{M6}
\alias{M6}
\title{National Saving (Year Ended March)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/economic-indicators/national-saving} for a full description.
}
\keyword{internal}
