% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{M7}
\alias{M7}
\title{Balance of Payments and International Investment Position}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/economic-indicators/balance-of-payments-and-international-investment-position} for a full description.
}
\keyword{internal}
