% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{R2}
\alias{R2}
\title{Reserve Bank Statistical Balance Sheet}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/reserve-bank/our-statistical-balance-sheet} for a full description.
}
\keyword{internal}
