% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{R3}
\alias{R3}
\title{Reserve Bank Analytical Accounts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/reserve-bank/our-analytical-accounts} for a full description.
}
\keyword{internal}
