% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBNZ.R
\docType{package}
\name{RBNZ}
\alias{RBNZ}
\title{Access Data from the Reserve Bank of New Zealand Website.}
\description{
Downloads data spreadsheets from the website of the Reserve Bank of New
Zealand (RBNZ) and reads them into R. See <https://www.rbnz.govt.nz/statistics>
for context.
}
\details{
The copyright notice for this data can be found at
         \url{https://www.rbnz.govt.nz/copyright}.
}
\author{
Jasper Watson
}
