% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S10}
\alias{S10}
\title{S10 - Banks: Balance sheet ($m)}
\arguments{
\item{option}{Allowed values are "default" (default) and "memo".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s10-banks-balance-sheet} for a full description.
}
