% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S32}
\alias{S32}
\title{Banks: Assets – Loans by Product}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s32-banks-assets-loans-by-product}
for a full description.
}
\keyword{internal}
