% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S33}
\alias{S33}
\title{S33 - Banks: Assets – Loans fully secured by residential mortgage by repricing ($m)}
\arguments{
\item{option}{Allowed values are "default" (default) and "longrun".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s33-banks-assets-loans-fully-secured-by-residential-mortgage-by-repricing} for a full description.
}
