% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S34}
\alias{S34}
\title{Banks: Assets – Loans and Repos by Industry}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s34-banks-assets-loans-by-industry}
for a full description.
}
\keyword{internal}
