% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S35}
\alias{S35}
\title{Banks: Assets – Loans by Business Size}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-assets-loans-by-business-size}
for a full description.
}
\keyword{internal}
