% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S36}
\alias{S36}
\title{Banks: Assets – Loans by Product}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-assets-business-loans-by-product}
for a full description.
}
\keyword{internal}
