% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S37}
\alias{S37}
\title{Banks: Assets – Agriculture Loans by Product}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-assets-agriculture-loans-by-product}
for a full description.
}
\keyword{internal}
