% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S40}
\alias{S40}
\title{S40 - Banks: Liabilities – Deposits by sector ($m)}
\arguments{
\item{option}{Allowed values are "default" (default) and "longrun".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s40-banks-liabilities-deposits-by-sector} for a full description.
}
