% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S50}
\alias{S50}
\title{Banks: Assets - Loans by Asset Quality}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default) and \dQuote{longrun}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-assets-loans-by-asset-quality}
for a full description.
}
\keyword{internal}
