% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T11}
\alias{T11}
\title{Savings Institutions: Balance Sheet}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/savings-institutions-balance-sheet} for a full description.
}
\keyword{internal}
