% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T21}
\alias{T21}
\title{Deposit Taking Finance Companies: Balance Sheet}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/deposit-taking-finance-companies-balance-sheet} for a full description.
}
\keyword{internal}
