% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T4}
\alias{T4}
\title{Non Banks: Funding and Claims by Sector}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/non-banks-funding-and-claims-by-sector} for a full description.
}
\keyword{internal}
