% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T40}
\alias{T40}
\title{Funds Under Management}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/managed-funds-funds-under-management} for a full description.
}
\keyword{internal}
