% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T41}
\alias{T41}
\title{Managed Fund Assets}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/managed-funds-assets-by-sector} for a full description.
}
\keyword{internal}
