% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T42}
\alias{T42}
\title{Life Insurance}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/life-insurance-assets-by-sector} for a full description.
}
\keyword{internal}
