% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T43}
\alias{T43}
\title{KiwiSaver}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/kiwisaver-assets-by-sector} for a full description.
}
\keyword{internal}
