% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T44}
\alias{T44}
\title{Other Registered Superannuation}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/other-registered-superannuation-funds-assets-by-sector} for a full description.
}
\keyword{internal}
