% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T45}
\alias{T45}
\title{Retail Unit Trusts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/retail-unit-trusts-assets-by-sector} for a full description.
}
\keyword{internal}
