% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T46}
\alias{T46}
\title{Cash Management Trusts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/cash-management-trusts-assets-by-sector} for a full description.
}
\keyword{internal}
