% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T47}
\alias{T47}
\title{Wholesale Trusts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/wholesale-trusts-assets-by-sector} for a full description.
}
\keyword{internal}
