% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{T48}
\alias{T48}
\title{Wholesale Trusts and Other Funds}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/non-banks-and-other-financial-institutions/wholesale-trusts-and-other-funds-assets-by-product-type} for a full description.
}
\keyword{internal}
