% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B1}
\alias{B1}
\title{Exchange Rates and TWI}
\arguments{
\item{option}{Allowed values are \dQuote{daily} (default) and \dQuote{monthly}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b1} for a full description.
}
\note{
The historical data is split across multiple files at source. We download
      all of them and merge together to form a complete history.
}
\keyword{internal}
