% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B2}
\alias{B2}
\title{Wholesale Interest Rates}
\arguments{
\item{option}{Allowed values are \dQuote{daily} (default), \dQuote{dailyclose},
\dQuote{monthly}, and \dQuote{monthlyclose}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b2} for a full description.
}
\note{
The historical data is split across multiple files at source. We download
      all of them and merge together to form a complete history.
}
\keyword{internal}
