% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B27}
\alias{B27}
\title{B27 - New interest bearing term PIE deposits interest rates}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b27-new-interest-bearing-term-pie-deposits-interest-rates}
for a full description.
}
\keyword{internal}
