% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{B4}
\alias{B4}
\title{Foreign Exchange Turnover}
\arguments{
\item{option}{Allowed values are \dQuote{NZD} (default), \dQuote{USD}
and \dQuote{EUR}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/b4} for a full description.
}
\keyword{internal}
