% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C12}
\alias{C12}
\title{Credit Card Balances}
\arguments{
\item{option}{Allowed values are \dQuote{monthend} (default) and
\dQuote{dailyaverage}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c12} for a full description.
}
\keyword{internal}
