% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C13}
\alias{C13}
\title{C13 - Credit card spending}
\arguments{
\item{option}{Allowed values are "issued" (default) and "used".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c13} for a full description.
}
\keyword{internal}
