% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C30}
\alias{C30}
\title{C30 - New residential mortgage lending by loan-to-valuation ratio (LVR) compliance data}
\arguments{
\item{option}{Allowed values are "" (default) and "".}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c30} for a full description.
}
\note{
The historical data is split across multiple files at source. We download all
      of them and merge together to form a complete history.
}
