% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C30}
\alias{C30}
\title{New Residential Mortgage Lending by Loan-To-Valuation Ratio}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default) and \dQuote{longrun}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c30} for a full description.
}
\note{
There are a few other files present on the webpage for this series but
      they are not downloaded. You will need to retrieve them manually if you
      want them.
}
\keyword{internal}
