% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C32}
\alias{C32}
\title{New and Existing Residential Mortgage Lending by Payment Type}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c32} for a full description.
}
\keyword{internal}
