% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C40}
\alias{C40}
\title{Residential Mortgage Lending by Debt-To-Income Purpose Use}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c40-residential-mortgage-lending-by-debt-to-income-dti-purpose-use}
for a full description.
}
\keyword{internal}
