% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C41}
\alias{C41}
\title{Residential Mortgage Borrower Gross Income (BGI)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c41-residential-mortgage-borrower-gross-income-bgi}
for a full description.
}
\keyword{internal}
