% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C50}
\alias{C50}
\title{Money and Credit Aggregates (Depository Corporations)}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default) and \dQuote{longrun}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c50-money-and-credit-aggregates}
for a full description.
}
\keyword{internal}
