% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C52}
\alias{C52}
\title{C52 - Depository corporations analytical accounts}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c52-depository-corporations-analytical-accounts} for a full description.
}
\keyword{internal}
