% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{C60}
\alias{C60}
\title{Credit Conditions Survey}
\description{
See \url{https://www.rbnz.govt.nz/statistics/c60-credit-conditions-survey}
for a full description.
}
\note{
There is a second dataset present (the ec60 file) which is not retrieved
      as it is in a non-standard format.
}
\keyword{internal}
