% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D10}
\alias{D10}
\title{Influences on Settlement Cash}
\description{
See \url{https://www.rbnz.govt.nz/statistics/d10} for a full description.
}
\note{
This data is in a non-standard format.
}
\keyword{internal}
