% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D30}
\alias{D30}
\title{Holdings of Central Government Debt Securities}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/new-zealand-debt-securities/holdings-of-central-government-debt-securities} for a full description.
}
\keyword{internal}
