% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D31}
\alias{D31}
\title{Non-Resident Holdings of Individual Bonds (Nominal Value)}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/new-zealand-debt-securities/non-resident-holdings-of-individual-bonds} for a full description.
}
\keyword{internal}
