% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{D9}
\alias{D9}
\title{Government Bond Turnover}
\arguments{
\item{option}{Allowed values are \dQuote{daily} (default) and \dQuote{monthly}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/d9} for a full description.
}
\keyword{internal}
