% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{M1}
\alias{M1}
\title{Prices}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default) and \dQuote{longrun}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/m1} for a full description.
}
\keyword{internal}
