% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBNZ.R
\name{RBNZ}
\alias{RBNZ}
\title{Access Data From the Reserve Bank of New Zealand Website.}
\description{
Downloads xlsx files from the website of the Reserve Bank of New Zealand (RBNZ) and reads them into R. All currently maintained datasets that are in the form of time series data can be retrieved. This includes the following series: B, C, R, S, L, J, T, D, M, and E2. See \url{https://www.rbnz.govt.nz/statistics} for more information.
}
\details{
The copyright notice for this data can be found at \url{https://www.rbnz.govt.nz/copyright}.
}
\author{
Jasper Watson
}
