% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeriesPage.R
\name{readSeriesPage}
\alias{readSeriesPage}
\title{Read a Series Webpage Into an XML object}
\usage{
readSeriesPage(series, quiet = TRUE)
}
\arguments{
\item{series}{Character string giving the name of the data series to retrieve.}

\item{quiet}{Logical to be passed to utils::download.file.}
}
\value{
An XML document returned by xml2::read_html containing the webpage
        pertaining to the desired series.
}
\description{
Read a Series Webpage Into an XML object
}
\author{
Jasper Watson
}
\keyword{internal}
