% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S10}
\alias{S10}
\title{Banks: Balance Sheet}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default), \dQuote{longrun}, and \dQuote{moredetail}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/series/registered-banks/banks-balance-sheet}
for a full description.
}
\keyword{internal}
