% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S30}
\alias{S30}
\title{Banks: Assets - Loans by Sector}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s30-banks-assets-loans-by-sector}
for a full description.
}
\keyword{internal}
