% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S31}
\alias{S31}
\title{Banks: Assets – Loans by Purpose}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s31-banks-assets-loans-by-purpose}
for a full description.
}
\keyword{internal}
