% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S40}
\alias{S40}
\title{Banks: Liabilities – Deposits by Sector}
\arguments{
\item{option}{Allowed values are \dQuote{default} (default) and \dQuote{longrun}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s40-banks-liabilities-deposits-by-sector}
for a full description.
}
\keyword{internal}
