% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{S41}
\alias{S41}
\title{Banks: Liabilities – Deposits by Industry}
\description{
See \url{https://www.rbnz.govt.nz/statistics/s41-banks-liabilities-deposits-by-industry}
for a full description.
}
\keyword{internal}
