% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesDescriptions.R
\name{SDDS}
\alias{SDDS}
\title{NZ's International Reserves and Foreign Currency Liquidity}
\arguments{
\item{option}{Allowed values are \dQuote{current} (default) and a year and month
specified in the form \dQuote{20nn-nn}.}
}
\description{
See \url{https://www.rbnz.govt.nz/statistics/sdds} for a full description.
}
\note{
This is a non-standard dataset.
}
\keyword{internal}
